refactor the trie-tree
