+++ operators/tokenizer/trie_tokenizer.hpp
@@ -14,66 +14,32 @@
 #include <optional>
 
 #include "unescape.h"
+#include "trietree.hpp"
 
 // This Trie Tree is C++ implementation of
 // https://github.com/BlinkDL/ChatRWKV/blob/main/rwkv_pip_package/src/rwkv/rwkv_tokenizer.py
 // Perf optimized by leveraging C++ features, but the algorithm is the same.
-class TrieTree {
+class RWKVTrieTree: public TrieTree {
  public:
   static constexpr int kMaxTokenLength_ = 128;
 
-  TrieTree(unsigned char ch = 0) : ch_(ch), to_(256) {}
+  RWKVTrieTree(unsigned char ch = 0) : TrieTree(ch){}
 
+  // keep the same function for source code understanding.
   void add(const std::string& key, int idx = 0,
            std::optional<int> value = std::optional<int>()) {
-    if (idx == key.length()) {
-      if (!value) {
-        value = key[0];
-      }
-      value_ = value;
-      return;
-    }
-
-    unsigned char ch = static_cast<unsigned char>(key[idx]);
-    if (to_[ch] == nullptr) {
-      to_[ch] = std::make_unique<TrieTree>(ch);
-    }
-    to_[ch]->add(key, idx + 1, value);
+            Add(key, idx, value); 
   }
 
   int find_longest(const std::string& key, size_t& idx) {
-    const TrieTree* u = this;
-    unsigned char ch = key[idx];
-
-    int tok_id = 0;
-    size_t idx_end = idx;
-    while (u->to_[ch]) {
-      u = u->to_[ch].get();
-      idx += 1;
-      if (u->value_) {
-        tok_id = *u->value_;
-        idx_end = idx;
-      }
-      if (idx == key.length()) {
-        break;
-      }
-      ch = key[idx];
-    }
-
-    idx = idx_end;
-    return tok_id;
+    return FindLongest(key, idx);
   }
-
- private:
-  std::vector<std::unique_ptr<TrieTree>> to_;
-  std::optional<int> value_;
-  unsigned char ch_;
 };
 
 class TrieTokenizer {
  private:
   std::map<int, std::string> idx2token;
-  TrieTree root;
+  RWKVTrieTree root;
 
  public:
   TrieTokenizer(const std::string& text_tokens) {
